LAYOUT: Upgrade for mantine v7
