+++ src/components/app-info/AppInfo.tsx
@@ -15,7 +15,7 @@ import { Link } from "react-router-dom";
 import { useCurrentUser } from "../../context/user.context";
 import { useMediaMatch } from "../../hooks/media-match";
 import { updateUserDetails } from "../../services/user.service";
-import { useNavBtnStyle } from "../../theme/modules/layout.styles";
+import classes from "../../theme/modules/Layout.module.css";
 import Changelog from "./Changelog";
 
 export default function AppInfo(
@@ -45,12 +45,10 @@ export default function AppInfo(
     if (userData?.seenChangelog === false) mutate({ seenChangelog: true });
   };
 
-  const { classes } = useNavBtnStyle({ active: false });
-
   return (
     <>
       {props.type === "text" && (
-        <Group {...props} spacing="xs" position={props.position ?? "center"}>
+        <Group {...props} gap="xs">
           <Text
             component={Link}
             to={userData ? "/about-app" : "/about"}
@@ -64,7 +62,7 @@ export default function AppInfo(
           <IconPoint size={12} />
           <Text
             td="underline"
-            sx={{ cursor: "pointer" }}
+            style={{ cursor: "pointer" }}
             fz="xs"
             color="dimmed"
             onClick={open}
